// types/employee.ts
export interface Employee {
  id: string;
  userId: string;
  name: string;
  email: string;
  phone: string;
  department: string;
  jobTitle: string;
  contractType: string;
  attendance: string;
  monthlySalary: number;
  workHours: number;
  age: number;
  DOB: Date;
  gender: string;
  address: string;
  city: string;
  state: string;
  startDate: Date;
}